tests/wpt/web-platform-tests/navigation-api/navigate-event/transitionWhile-history-pushState.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  let start_length = history.length;
  navigation.onnavigate = t.step_func(e => {
    e.transitionWhile(new Promise(resolve => t.step_timeout(() => {
      resolve();
      t.step_timeout(t.step_func_done(() => {
        assert_equals(location.hash, "#1");
        assert_equals(history.state, "update");
        assert_equals(history.length, start_length + 1);
      }, 0));
    }, 0)));
  });

  history.pushState("update", "", "#1");
  assert_equals(location.hash, "#1");
  assert_equals(history.state, "update");
  assert_equals(history.length, start_length + 1);
}, "event.transitionWhile() should proceed if the given promise resolves");
</script>
